+++ Assets/Scripts/Client/UI/Panels/Battle/Action Bars/ActionBar.cs
@@ -21,7 +21,7 @@ public class ActionBar : UIContainer
         base.Deinitialize();
     }
 
-    public override void DoUpdate(int deltaTime)
+    public override void DoUpdate(float deltaTime)
     {
         base.DoUpdate(deltaTime);
